# app.py
from flask import Flask, request, redirect, jsonify, abort, render_template, make_response

import json
import datetime
from nanoid import generate

from redis_om import Migrator
from redis_om.model import NotFoundError
from pydantic import ValidationError

from src.schema import Url
from src.logger import initLogger
from src.constant import *
from src.utils import *
from src.swagger import initSwaggerUI

ROUTE_NAME = ('shorten', 'swagger', )

app = Flask(__name__)
initSwaggerUI(app)
logger = initLogger()

@app.route('/')
@app.route('/shorten', methods=['GET'])
def index():
    return render_template('index.html')

def validateShortkey(shortKey):
    if len(shortKey) > ALIAS_URL_MAX_LENGTH:
        return f'Bad request: The character length of the alias cannot exceed {ALIAS_URL_MAX_LENGTH}.'
    elif len(shortKey) < ALIAS_URL_MIN_LENGTH:
        return f'Bad request: The character length of the alias cannot be less than {ALIAS_URL_MIN_LENGTH}.'
    if shortKey in ROUTE_NAME:
        return f'Bad request: the short key cannot be the same as route name.'
    try:
        validateCharacters(shortKey)
    except Exception as e:
        return f'Bad request: {e}'

def isShortkeyExist(shortKey):
    # check the short URL in use or not
    results = find_by_shortkey(shortKey)
    return len(results) > 0

@app.route('/api/service/admin/<user>', methods=['GET'])
def checkSystemAdmin(user):
    if user in ADMIN_USER_ID:
        return jsonify('ok')
    return abort(make_response(jsonify(message=f'User "{user}" is not admin.'), 400))

@app.route('/<shortKey>', methods=['GET'])
@logExecutionTime
def redirectUrl(shortKey):
    error = validateShortkey(shortKey)
    if error is not None:
        return abort(make_response(jsonify(message=error), 400))

    # Check if the response is already cached
    results = find_by_shortkey(shortKey)
    if len(results) == 0:
        shortUrl = SHORT_KEY_PREFIX + shortKey
        error = f'Not found original url by {shortUrl}'
        return abort(make_response(jsonify(message=error), 404))

    url = results[0]
    longUrl = url.original_url
    url.utilization = url.utilization + 1
    url.save()
    return redirect(longUrl)

@app.route('/api/shorten', methods=['POST'])
@logExecutionTime
def shortenUrl():
    try:
        data = json.loads(request.data.decode())
        userId = data['user_id']
        originalUrl = data['original_url']
        shortKey = data['short_key']
        expireType = data['expire_type']
    except:
        return abort(make_response(jsonify(message='Bad request'), 400))

    # validate for long url
    try:
        validateUrl(originalUrl)
    except Exception as e:
        error = f'Bad request: {e}'
        return abort(make_response(jsonify(message=error), 400))
    
    # validate for short key
    if len(shortKey) > 0:
        error = validateShortkey(shortKey)
        if error is not None:
            return abort(make_response(jsonify(message=error), 400))
        if isShortkeyExist(shortKey):
            error = f'Bad request: The short key "{shortKey}" is in use.'
            return abort(make_response(jsonify(message=error), 400))

    # validate for expire type
    if expireType not in ('1month', '3month', '6month', '1year', 'indefinitely'):
        error = 'Bad request: invalid expire time'
        return abort(make_response(jsonify(message=error), 400))
    expireTime = type2date(expireType)

    try:
        shortUrl, urlId = generate_shorturl(userId, originalUrl, shortKey, expireTime)
        warning = checkBroadcomUrlAccess(originalUrl)
        return jsonify(short_url=shortUrl, long_url=originalUrl, url_id=urlId, warning=warning)
    except ValidationError as e:
        error = 'Validation error: ' + str(e)
        return abort(make_response(jsonify(message=error), 500))
    except Exception as e:
        error = 'Internal error: ' + str(e)
        return abort(make_response(jsonify(message=error), 500))
   
@app.route('/api/url/<id>', methods=['GET'])
def findById(id):
    try:
        url = Url.get(id)
        result = url.dict()
        result.update({'short_url': SHORT_KEY_PREFIX + result['short_key']})
        return result
    except NotFoundError:
        error = 'Not found by url id: ' + id
        return abort(make_response(jsonify(message=error), 404))

@app.route('/api/url/<id>', methods=['DELETE'])
def deleteById(id):
    try:
        ret = Url.delete(id)
        # Delete returns 1 if the url existed and was deleted, or 0 if they didn't exist.
        app.logger.info(f'delete url {id}: {ret}')
        return jsonify('ok')
    except NotFoundError:
        error = 'Not found by url id: ' + id
        return abort(make_response(jsonify(message=error), 404))

@app.route('/api/url/<id>', methods=['POST'])
def updateById(id):
    try:
        url = Url.get(id)
    except NotFoundError:
        error = 'Not found by url id: ' + id
        return abort(make_response(jsonify(message=error), 404))
    
    try:
        data = json.loads(request.data.decode())
        originalUrl = data['original_url']
        validateUrl(originalUrl)
    except Exception as e:
        error = f'Bad request: {e}'
        return abort(make_response(jsonify(message=error), 400))
    
    try:
        url.original_url = originalUrl
        url.hash_original = url2hash(originalUrl)
        url.save()
        shortUrl = SHORT_KEY_PREFIX + url.short_key
        return jsonify(short_url=shortUrl, long_url=url.original_url, url_id=url.pk)
    except ValidationError as e:
        error = 'Validation error: ' + str(e)
        app.logger.error(error)
        return abort(make_response(jsonify(message=error), 500))
    except Exception as e:
        error = 'Internal error: ' + str(e)
        return abort(make_response(jsonify(message=error), 500))
    
@app.route('/api/longurl', methods=['GET'])
def queryByLongurl():
    try:
        data = json.loads(request.data.decode())
        longUrl = data['original_url']
        if len(longUrl) == 0:
            return abort(400, f'Bad request: No long url specified.')
    except:
        return abort(make_response(jsonify(message='Bad request'), 400))

    hash_value = url2hash(longUrl)
    try:
        datas = Url.find(
            (Url.hash_original == hash_value)
        ).all()
    except NotFoundError:
        error = 'Not found by long url: ' + longUrl
        return abort(make_response(jsonify(message=error), 404))

    app.logger.debug(f'Hash value "{hash_value}" found {len(datas)} records')
    return build_results(datas)

@app.route('/api/longurl', methods=['POST'])
@logExecutionTime
def updateLongurlByShortkey():
    try:
        data = json.loads(request.data.decode())
        shortKey = data['short_key']
        originalUrl = data['original_url']
    except:
        return abort(make_response(jsonify(message='Bad request'), 400))

    # validate for short key
    error = validateShortkey(shortKey)
    if error is not None:
        return abort(make_response(jsonify(message=error), 400))
    urls = find_by_shortkey(shortKey)
    if len(urls) == 0:
        error = 'Not found by short key: ' + shortKey
        return abort(make_response(jsonify(message=error), 404))

    # validate for long url
    try:
        validateUrl(originalUrl)
    except Exception as e:
        error = f'Bad request: {e}'
        return abort(make_response(jsonify(message=error), 400))

    try:
        url = urls[0]
        app.logger.debug(url)
        url.original_url = originalUrl
        url.hash_original = url2hash(originalUrl)
        url.save()
        shortUrl = SHORT_KEY_PREFIX + url.short_key
        warning = checkBroadcomUrlAccess(originalUrl)
        return jsonify(short_url=shortUrl, long_url=url.original_url, url_id=url.pk, warning=warning)
    except ValidationError as e:
        error = 'Validation error: ' + str(e)
        app.logger.error(error)
        return abort(make_response(jsonify(message=error), 500))
    except Exception as e:
        error = 'Internal error: ' + str(e)
        return abort(make_response(jsonify(message=error), 500))

@app.route('/api/shortkey/<shortKey>', methods=['GET'])
def queryByShortkey(shortKey):
    # validate for short key
    error = validateShortkey(shortKey)
    if error is not None:
        return abort(make_response(jsonify(message=error), 400))
    urls = find_by_shortkey(shortKey)
    if len(urls) == 0:
        error = 'Not found by short key: ' + shortKey
        return abort(make_response(jsonify(message=error), 404))
    return build_results(urls)

@app.route('/api/shortkey/<id>/<shortKey>', methods=['POST'])
@logExecutionTime
def updateShortkeyById(id, shortKey):
    # validate for short key
    error = validateShortkey(shortKey)
    if error is not None:
        return abort(make_response(jsonify(message=error), 400))
    if isShortkeyExist(shortKey):
        error = f'Bad request: The short key "{shortKey}" is in use.'
        return abort(make_response(jsonify(message=error), 400))

    try:
        url = Url.get(id)
        app.logger.debug(url)
        url.short_key = shortKey
        url.save()
        shortUrl = SHORT_KEY_PREFIX + url.short_key
        return jsonify(short_url=shortUrl, long_url=url.original_url, url_id=url.pk)
    except ValidationError as e:
        error = 'Validation error: ' + str(e)
        app.logger.error(error)
        return abort(make_response(jsonify(message=error), 500))
    except NotFoundError:
        error = 'Not found by url id: ' + id
        return abort(make_response(jsonify(message=error), 404))

@app.route('/api/urls', methods=['GET'])
def queryByUser():
    user = request.args.get('user', '')
    if len(user) == 0:
        return abort(400, f'Bad request: No user specified.')

    datas = Url.find(
        (Url.user_id == user)
    ).all()
    datas.sort(reverse=True)
    app.logger.info(f'User "{user}" has {len(datas)} urls')
    return build_results(datas)

def build_results(urls):
    response = []
    for url in urls:
        data = url.dict()
        data.update({'short_url': SHORT_KEY_PREFIX + data['short_key']})
        response.append(data)
    return {"results": response}

def find_by_shortkey(shortKey):
    try:
        datas = Url.find(
            (Url.short_key == shortKey)
        ).all()
    except NotFoundError:
        datas = []
    app.logger.info(f'Short key "{shortKey}" found {len(datas)} records')
    return datas

def create_url(short_key, long_url, expire_time, user_id):
    now = datetime.date.today()
    hash_value = url2hash(long_url)
    new_url = Url(original_url=long_url, hash_original=hash_value, short_key=short_key, 
                  expire_time=expire_time, create_at=now, user_id=user_id)
    new_url.save()
    if expire_time is not None:
        delta_second = (expire_time - now).days * 24 * 3600
        app.logger.info(f'Set expire time {delta_second} seconds for short key {short_key}')
        new_url.expire(delta_second)
    return new_url.pk

def generate_shorturl(user_id, original_url, short_key, expire_time):
    if len(short_key) > 0:
        pk = create_url(short_key, original_url, expire_time, user_id)
        short_url = SHORT_KEY_PREFIX + short_key
        return short_url, pk
    
    for _ in range(SHORT_KEY_GENERATION_RETRIES_COUNT):
        short_key = generate(size=SHORT_URL_LENGTH)
        # check the short URL in use or not
        results = find_by_shortkey(short_key)
        if len(results) == 0:
            pk = create_url(short_key, original_url, expire_time, user_id)
            short_url = SHORT_KEY_PREFIX + short_key
            return short_url, pk

    raise Exception('Fail to generate a unique short url')

# Create a RedisSearch index for instances of the Url model.
Migrator().run()

app.logger.info(f"Connect to redis uri: {REDIS_OM_URL}")