# doc/swagger.yml
swagger: '2.0'
info:
  title: URL Shortener API
  description: This is the document of vSAN URL Shortener restful API.
  version: 1.0.0
  contact: https://chat.google.com/room/AAAA1p6A3v8?cls=7
host: vsanvia.vmware.com
schemes:
  - http
servers:
  - url: http://vsanvia.vmware.com/
tags:
  - name: public
    description: Public API for users
  - name: private
    description: Private API for admins
definitions:
  ShortenReqData:
    properties:
      short_key:
        example: HELLO-url
        type: string
      original_url:
        example: https://confluence.eng.vmware.com/pages/viewpage.action?spaceKey=SABU&title=vSAN+Slackbot+Notification+Service
        type: string
      expire_type:
        example: 6month
        type: string
      user_id:
        example: zhouhuang
        type: string
  UpdateLongUrlReqData:
    properties:
      short_key:
        example: HELLO-url
        type: string
      original_url:
        example: https://confluence.eng.vmware.com/display/SABU/vSAN+Short+URL+Service
        type: string
  QueryLongUrlReqData:
    properties:
      original_url:
        example: https://confluence.eng.vmware.com/display/SABU/vSAN+Short+URL+Service
        type: string
  ShortUrl:
    properties:
      short_url:
        example: https://vsanvia.vmware.com/test-url
        type: string
      long_url:
        example: https://confluence.eng.vmware.com/pages/viewpage.action?spaceKey=SABU&title=vSAN+Slackbot+Notification+Service
        type: string
      url_id:
        example: 01J3MFXQB7NPF7WVQ6CFTB4Z7W
        type: string
  UrlDetail:
    properties:
      short_url:
        example: https://vsanvia.vmware.com/test-url
        type: string
      original_url:
        example: https://confluence.eng.vmware.com/pages/viewpage.action?spaceKey=SABU&title=vSAN+Slackbot+Notification+Service
        type: string
      hash_original:
        example: fab8d5cba0ca835fcc642eb8c093cea0e2244e03ba08ee67d0121631ce7c5fac
        type: string
      create_at:
        example: Thu, 25 Jul 2024 00:00:00 GMT
        type: string
      expire_time:
        example: Sat, 25 Jan 2025 00:00:00 GMT
        type: string
      user_id:
        example: zhouhuang
        type: string
      pk:
        example: 01J3MFXQB7NPF7WVQ6CFTB4Z7W
        type: string
  UrlDetailList:
    properties:
      results:
        type: array
        items:
          $ref: '#/definitions/UrlDetail'
  HTTPError:
    properties:
      message:
        example: error message
        type: string
    type: object
paths:
  /{shortKey}:
    get:
      summary: Redirect to long url by short key
      tags:
        - public
      parameters:
        - in: path
          name: shortKey
          required: true
          type: string
          example: HELLO-web
          description: The short key used to query in db
      responses:
        302:
          description: Redirect to long url.
          headers:
            Location:
              description: The URL to which the client should be redirected.
              schema:
                type: string
                example: /new-path
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "This resource has temporarily moved to /new-path."
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/HTTPError'
        404:
          description: Not found
          schema:
            $ref: '#/definitions/HTTPError'
  /api/shorten:
    post:
      summary: Create a short url
      tags:
        - public
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: shorten url info
          required: true
          description: The short url info to be created
          schema:
            $ref: '#/definitions/ShortenReqData'
      responses:
        200:
          description: The info of created short url
          schema:
            $ref: '#/definitions/ShortUrl'
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTPError'
        500:
          description: Internal Error
          schema:
            $ref: '#/definitions/HTTPError'
  /api/url/{id}:
    get:
      summary: Fetch one short url detail by id
      tags:
        - public
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          type: string
          example: 01J3MFXQB7NPF7WVQ6CFTB4Z7W
          description: The uuid of short url
      responses:
        200:
          description: The details of short url
          schema:
            $ref: '#/definitions/UrlDetail'
        404:
          description: Not found
          schema:
            $ref: '#/definitions/HTTPError'
    delete:
      summary: Delete one short url by id
      tags:
        - public
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          type: string
          example: 01J3MFXQB7NPF7WVQ6CFTB4Z7W
          description: The uuid of short url
      responses:
        200:
          description: The short url deleted.
          schema:
            type: string
            example: ok
        404:
          description: Not found
          schema:
            $ref: '#/definitions/HTTPError'
  /api/longurl:
    get:
      summary: Query url infos by long url
      tags:
        - private
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: long url
          required: true
          description: The long url used to query in db
          schema:
            $ref: '#/definitions/QueryLongUrlReqData'
      responses:
        200:
          description: The url info list queried by long url
          schema:
            $ref: '#/definitions/UrlDetailList'
        404:
          description: Not found
          schema:
            $ref: '#/definitions/HTTPError'
    post:
      summary: Update the long url by short key
      tags:
        - public
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: update long url info
          required: true
          description: The long url info to be updated
          schema:
            $ref: '#/definitions/UpdateLongUrlReqData'
      responses:
        200:
          description: The info of updated short url
          schema:
            $ref: '#/definitions/ShortUrl'
        400:
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTPError'
        404:
          description: Not found
          schema:
            $ref: '#/definitions/HTTPError'
        500:
          description: Internal Error
          schema:
            $ref: '#/definitions/HTTPError'
  /api/shortkey/{shortKey}:
    get:
      summary: Query url infos by short key
      tags:
        - public
      produces:
        - application/json
      parameters:
        - in: path
          name: shortKey
          required: true
          type: string
          example: HELLO-web
          description: The short key used to query in db
      responses:
        200:
          description: The url info list queried by short key (The number of url info list should be one.)
          schema:
            $ref: '#/definitions/UrlDetailList'
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/HTTPError'
        404:
          description: Not found
          schema:
            $ref: '#/definitions/HTTPError'
  /api/shortkey/{id}/{shortKey}:
    post:
      summary: Update the short key by url id
      tags:
        - public
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          type: string
          example: 01J3MFXQB7NPF7WVQ6CFTB4Z7W
          description: The uuid of short url
        - in: path
          name: shortKey
          required: true
          type: string
          example: HELLO-web
          description: The new short key to be updated
      responses:
        200:
          description: The info of updated short url
          schema:
            $ref: '#/definitions/ShortUrl'
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/HTTPError'
        404:
          description: Not found
          schema:
            $ref: '#/definitions/HTTPError'
        500:
          description: Internal Error
          schema:
            $ref: '#/definitions/HTTPError'
  /api/urls:
    get:
      summary: Query url infos by user name
      tags:
        - private
      produces:
        - application/json
      parameters:
        - in: query
          name: user
          required: true
          type: string
          example: zhouhuang
          description: VMware AD account
      responses:
        200:
          description: The url info list queried by user name
          schema:
            $ref: '#/definitions/UrlDetailList'
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/HTTPError'